Further update on SingleStageSelect() graphics and ui. Title is changed from test_title.png to test_Planet.png and test_Escape.png so that later can be modified.

+++ Java_Gaming/src/main/model/Position.java
@@ -1,12 +1,12 @@
 package main.model;
 
 public class Position {
-	private int x, y;
+	private double x, y;
 	public void setX(int x) {this.x = x;}
 	public void setY(int y) {this.y = y;}
-	public int getX() {return x;}
-	public int getY() {return y;}
-	public Position(int x, int y) {
+	public double getX() {return x;}
+	public double getY() {return y;}
+	public Position(double x, double y) {
 		this.x = x; this.y = y;
 	}
 }
